src/7- Image Similarity/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
timm>=0.6.0
matplotlib>=3.4.0 